job pool merge
